Add health check endpoint + some cleanup
