+++ config/routes.rb
@@ -2,5 +2,6 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   namespace :api do
     post "/graphql", to: "graphql#execute"
+    get "/health", to: "health#index"
   end
 end
